Add files via upload
The bionic animation "Life" was created in 1970 by the British mathematician J.H.
Conway invented. We consider a two-dimensional matrix with n * m
Elements (cells) of type char. If a cell contains a space,
then she is "dead". It is a "living" cell if one is in it
"*" is entered. Several neighboring living cells form one
Organism.
A cell is surrounded by 8 neighboring cells when they are not straight
are at the edge of the matrix. To the shape of an organism in the
To determine the next generation, the following rules apply, which apply to the
entire matrix are to be used:
1. If a living cell has 0, 1 or more than 3 living cells
Neighboring cells, she dies of loneliness or overpopulation.
2. If a living cell has 2 or 3 living neighboring cells, then
it will continue to exist in the next generation.
3. If a dead cell is surrounded by exactly 3 living cells, then
a new living cell emerges from it.
Write a program that will help you develop one
Understand the organism on the screen.
Step 1: Depending on the user's wishes, the starting state of the matrix is
is loaded from a file or is generated by a random number generator
Start state calculated. The user can decide whether a
step-by-step processing (development) or a "flowing animation"
it is asked for.
Step 2: The functionality for grade 3 has been implemented. The user
You can choose between at least 3 files (i.e. at least 3
Select start states). When generating a start state by
the user can specify the random number generator to what percentage
the playing field is to be filled with living cells.
Step 3: The functionality for grade 2 has been implemented. The user
can edit start states in the work field and then under a
Save the file name of your choice.